refactor: move nativeCurrency logic to a hook
