+++ ui/components/app/assets/defi-list/defi-tab.tsx
@@ -3,9 +3,7 @@ import { AssetListProps } from '../asset-list/asset-list';
 import AssetListControlBar from '../asset-list/asset-list-control-bar';
 import DefiList from './defi-list';
 
-export default function DeFiTab({
-  onClickAsset,
-}: Omit<AssetListProps, 'getNativeCurrencySymbol'>) {
+export default function DeFiTab({ onClickAsset }: AssetListProps) {
   return (
     <>
       <AssetListControlBar showImportTokenButton={false} />
